Do not enqueue multiple GET_LIGHT_STATE for all bulbs on a single gateway
lgtd_gateway_handle_light_status is called per bulb and not per gateway
and we trigger the refresh of all the bulbs on the gateway from this
function.

This means that we could potentially enqueue multiple GET_LIGHT_STATE
packets for all bulbs on the same gateway within a single libevent
iteration.

With this changeset each gateway now keep track of when the last
GET_LIGHT_STATE for all bulbs was sent. If it's more recent than when
the last packet was sent then we know that one has already been
enqueued.