Make the pipe/socket read loops fully resilient against incomplete json
In other words you can send your request character by character with as
much as white space as you want (and as allowed by JSON) and lightsd and
it will still sort it out.

Among some rework in the socket read loop one big fix was to call
jsmn_init before each call to jsmn_parse. So not sure if jsmn is really
working as explained in its documentation.

This changeset also moves -DJSMN_STRICT=1 to the top-level
CMakeLists.txt so that the tests and lightsd use the same code.