Handle arbitrary large JSON-RPC request up to 4KiB
Turns out it's pretty easy to run out of jsmn tokens, so this changeset
uses reallocarray to allocate them. This means that we have to parse the
request twice (one time to check how many tokens we need, one time to do
the actual parsing).

Since tokens are now allocated dynamically it makes a lot more sense to
use jsmn's PARENT_LINKS and this changeset does that.