Fix command pipe support & cleanup of pipes/sockets
lightsc.sh is now configured at build time, packages will work out of
the box.

Most importantly the user/group of the parent directory of any socket or
command pipe is now properly set when the -u option is used. This means
that command pipes work again and that both pipes and sockets can be
properly cleaned-up at exit.